// pwa/src/api/models/bookJsonld.ts
/**
 * Generated by orval v8.0.0-rc.0 🍺
 * Do not edit manually.
 * Hello API Platform
 * OpenAPI spec version: 1.0.0
 */
import type { BookJsonldContext } from "./bookJsonldContext";
import type { BookJsonldRating } from "./bookJsonldRating";
import type { BookJsonldBody } from "./bookJsonldBody";

export interface BookJsonld {
  readonly "@context"?: BookJsonldContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
  rating?: BookJsonldRating;
  body?: BookJsonldBody;
  author?: string;
  reviews?: string[];
}